@pedroigor 2022-04-07
I always thought you can not store engine instances. Indeed a nice improvement