@mposolda 2022-04-08
> Does it make sense to add an options to disable/enable caching?
> 
> The default would be enabled.

@pedroigor I was actually thinking about this as well :-) I've added it and marked to enabled by default.